order placed by user

+++ Domain/Entities/OrderItem.cs
@@ -12,8 +12,8 @@ namespace Domain.Entities
         public int OrderId { get; set; } // Foreign key to Order
         public int ProductId { get; set; } // Foreign key to Product
         public int Quantity { get; set; } // Quantity of the product in the order
-        public decimal UnitPrice { get; set; } // Price per unit at the time of order
-        public decimal TotalPrice => UnitPrice * Quantity; // Total price for this order item
+        public double UnitPrice { get; set; } // Price per unit at the time of order
+        public double TotalPrice => UnitPrice * Quantity; // Total price for this order item
 
         public Order Order { get; set; } // Navigation property to Order entity
         public Product Product { get; set; } // Navigation property to Product entity
